refactor: extract workspace state module
- Move workspace state derivation from cli.ts to module
- Extract effective config and local override application
- Add colocated tests for workspace state helpers

Made-with: Cursor
